Update PR dossier.html
Adding draggable="true" to each section allows the window to be selected and dragged, but not dropped, and nothing in the section can be selected. But without draggable="true", the sections can't be selected or dragged or dropped, but the contents can be selected. If the sections start out as display:none; they will not display, even when the folder is clicked. It's almost as if the script doesn't run at all, although it's copied and pasted from the undergroundfm page, with the id's changed for the DOM ID and Winamp and Winamp skins removed from the script. I've tried all sorts of different CSS thinking maybe it was a default attribute (like display:block; or display:inline;) throwing it off and breaking the script, but I'm out of ideas on it. New problem: clicked anywhere or anything on the entire webpage causes the mugshot section or the profile section or both to disappear and nothing brings them back.